fix: cleanup mock objects from blueprint detail
